requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.9.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
Django==5.2.5
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
joblib==1.5.3
multidict==6.7.0
mysqlclient==2.2.7
nltk==3.9.2
numpy==2.4.0
openai==0.28.0
openrouter==0.1.1
packaging==25.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.16.3
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
yarl==1.22.0
